# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0", 
      "title": "InstaLearn Project",
      "description": "InstaLearn Application API",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:3000",
    "basePath": "/",
    "tags": [
      {
        "name": "Autenticazione",
        "description": "API per l'iscrizione e autenticazione al sito"
      },
      {
        "name":"Insegnanti",
        "description":"API per la gestione degli insegnanti"
      },
      {
        "name":"Test",
        "description":"API per la gestione dei test"
      },
      {
        "name":"Materie",
        "description":"API per la gestione delle materie"
      },
      {
        "name":"Utenti",
        "description":"API per la gestione degli utenti"
      }
    ],
    "paths": {
        "/signup": {
            "post": {
                "tags": [
                    "Autenticazione"
                ],
                "summary": "Permette di registrarsi al sito",
                "parameters":[{
                    "in": "body",
                    "name": "utente",
                    "description": "Utente da registrare",
                    "required":"true",
                    "schema":
                    {
                        "$ref":"#/definitions/User"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400":{
                        "description": "L'utente è già presente nel sistema"
                    }
                }
            }
        },
        "/login":
        {
            "post": {
                "tags": [
                    "Autenticazione"
                ],
                "summary": "Permette di registrarsi al sito",
                "parameters":[{
                    "in": "body",
                    "name": "CredenzialiUtente",
                    "description": "Credenziali dell'utente che effettua il login",
                    "required":"true",
                    "schema":
                    {
                        "$ref":"#/definitions/UserCredentials"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK, restituisce l'id dell'utente per creare la sessione",
                        "schema":
                        {
                            "schema":{"$ref":"#/definitions/IdUtente"}
                        }
                        
                    },
                    "400":{
                        "description": "L'utente non è presente nel sistema"
                    }
                }
            }
        },
        "/cercaInsegnante":
        {
            "post": {
                "tags": [
                    "Insegnanti"
                ],
                "summary": "Permette di ricercare un insegnante in base ad un topic e a ad un livello di abilità",
                "parameters":[{
                    "in": "body",
                    "name": "TopicSkillUtente",
                    "description": "Un Json composto da un id topic, un livello di skill e un id utente",
                    "required":"true",
                    "schema":
                    {
                        "$ref":"#/definitions/TopicSkillUtente"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK, restituisce un Array di Insegnanti disponibili",
                        "schema":
                        {
                            "$ref":"#/definitions/ArrayDiInsegnanti"
                        }
                        
                    },
                    "400":{
                        "description": "L'utente non è presente nel sistema"
                    }
                }
            }
        },
        "/generaTest":
        {
            "post": {
                "tags": [
                    "Test"
                ],
                "summary": "Permette di creare un test",
                "parameters":[{
                    "in": "body",
                    "name": "TopicSkillUtente",
                    "description": "Un Json composto da un id topic, un livello di skill e un id utente (non obbligatorio)",
                    "required":"true",
                    "schema":
                    {
                        "$ref":"#/definitions/TopicSkillUtente"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK, restituisce un Array di Domande",

                        "schema":
                        {
                            "$ref":"#/definitions/ArrayDiDomande"
                        }
                        
                    },
                    "400":{
                        "description": "Non esiste un test componibile con tali informazioni"
                    }
                }
            }
        },
        "/getTestDisponibiliPerUtente":
        {
            "post": {
                "tags": [
                    "Test"
                ],
                "summary": "Permette di vedere che test sono diponibili",
                "parameters":[{
                    "in": "body",
                    "name": "IdUtente",
                    "description": "Id dell'utente il quale vuole sapere i quiz per lui disponibili",
                    "required":"true",
                    "schema":
                    {
                        "$ref":"#/definitions/IdUtente"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK, restituisce un Array contente le informazioni base dei Test consigliati",
                        "schema":
                        {
                            "$ref":"#/definitions/ArrayDiBaseTest"
                        }
                    },
                    "400":{
                        "description": "Non esiste un test componibile con tali informazioni"
                    }
                }
            }
        },
        "/getMateria":
        {
            "post": {
                "tags": [
                    "Materie"
                ],
                "summary": "Permette di sapere le informazioni su una materia a partire dall'id",
                "parameters":[{
                    "in": "body",
                    "name": "IdMateria",
                    "description": "Id della materia che si vuole cercare",
                    "required":"true",
                    "schema":
                    {
                        "$ref":"#/definitions/IdMateria"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK, restituisce le informazioni sulla materia",
                        "schema":
                        {
                            "$ref":"#/definitions/Materia"
                        }
                    },
                    "400":{
                        "description": "Non esiste una materia componibile con tali informazioni"
                    }
                }
            }
        },
        "/getDatiUtente":
        {
            "post": {
                "tags": [
                    "Utenti"
                ],
                "summary": "Permette di sapere le informazioni di un utente tramite il suo username",
                "parameters":[{
                    "in": "body",
                    "name": "Username",
                    "description": "Username dell'utente di cui si vogliono sapere le informazioni",
                    "required":"true",
                    "schema":
                    {
                        "$ref":"#/definitions/Username"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK, restituisce le informazioni sull'utente",
                        "schema":
                        {
                            "$ref":"#/definitions/UserPublic"
                        }
                    },
                    "400":{
                        "description": "Non esiste un utente componibile con tali informazioni"
                    }
                }
            }
        },
        "/getDatiSeStesso":
        {
            "post": {
                "tags": [
                    "Utenti"
                ],
                "summary": "Permette di sapere le proprie informazioni, per farlo è necessario il token (id utente)",
                "parameters":[{
                    "in": "body",
                    "name": "IdUtente",
                    "description": "Id dell'utente che vuole sapere le proprie informazioni",
                    "required":"true",
                    "schema":
                    {
                        "$ref":"#/definitions/IdUtente"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "OK, restituisce le informazioni private sull'utente",
                        "schema":
                        {
                            "$ref":"#/definitions/UserPrivate"
                        }
                    },
                    "400":{
                        "description": "Non esiste un utente componibile con tali informazioni"
                    }
                }
            }
        },
        "/getMaterie":
        {
            "post": {
                "tags": [
                    "Materie"
                ],
                "summary": "Permette di conoscere tutte le materie presenti",
                "responses": {
                    "200": {
                        "description": "OK, restituisce le materie",
                        "schema":
                        {
                            "$ref":"#/definitions/ArrayDiMaterie"
                        }
                    },
                    "400":{
                        "description": "Non esistono materie"
                    }
                }
            }
        }
    },
    "definitions": {
        "User": {
            "required": [
                "Email",
                "Password",
                "Nome",
                "Cognome",
                "Username"
            ],
            "properties": {
                "Email": {
                    "type": "string"
                },
                "Password":
                {
                    "type":"string"
                },
                "Nome": {
                    "type": "string"
                },
                "Cognome": {
                    "type": "string"
                },
                "Username":{"type":"string"}
            }
        },
        "UserPublic": {
            "required": [
                "Nome",
                "Cognome",
                "Username",
                "Descrizione",
                "Skills"
            ],
            "properties": {
                "Nome": {
                    "type": "string"
                },
                "Cognome": {
                    "type": "string"
                },
                "Descrizione":{"type":"string"},
                "Username":{"type":"string"},
                "Skills":{
                    "$ref":"#/definitions/ArrayDiTopicSkill"
                }
            }
        },
        "UserPrivate": {
            "required": [
                "Nome",
                "Cognome",
                "Username",
                "Descrizione",
                "Email",
                "Skills"
            ],
            "properties": {
                "Nome": {
                    "type": "string"
                },
                "Cognome": {
                    "type": "string"
                },
                "Email":{"type":"string"},
                "Descrizione":{"type":"string"},
                "Username":{"type":"string"},
                "Skills":{
                    "$ref":"#/definitions/ArrayDiTopicSkill"
                }
            }
        },
        "Username":
        {
            "required":["Username"],
            "properties":{"Username":{"type":"string"}}
        },
        "UserCredentials": {
            "required": [
                "Email",
                "Password"
            ],
            "properties": {
                "Email": {
                    "type": "string"
                },
                "Password":
                {
                    "type":"string"
                }
            }
        },
        "IdUtente":
        {
            "required":["Id"],
            "properties":{
                "Id":{"type":"string"}
            }
        },
        "TopicSkillUtente":
        {
            "required":["Id_topic","Skill"],
            "properties":{
                "Id_topic":{"type":"string"},
                "Skill":{"type":"integer"},
                "Id":{"type":"string"},
                "Materia":{"type":"string"}
            }
        },
        "Insegnante":
        {
            "required":["Skill","Username"],
            "properties":{
                "Skill":{"type":"integer"},
                "Username":{"type":"string"}
            }
        },
        "ArrayDiInsegnanti":
        {
            "type":"array",
            "items":
            {
                "$ref":"#/definitions/Insegnante"
            }
        },
        "Opzione":
        {
            "required":["Testo","Id_opzione"],
            "properties":
            {
                "Id_opzione":{"type":"string"},
                "Testo":{"type":"string"}
            }
        },
        "ArrayDiOpzioni":
        {
            "type":"array",
            "items":
            {
                "$ref":"#/definitions/Opzione"
            }
        },
        "Domanda":
        {
            "required":["Id_domanda","Testo","Opzioni"],
            "properties":
            {
                "Id_domanda":{"type":"string"},
                "Testo":{"type":"string"},
                "Opzioni":{"$ref":"#/definitions/ArrayDiOpzioni"}
            }
        },
        "ArrayDiDomande":
        {
            "type":"array",
            
            "items":
            {
                "$ref":"#/definitions/Domanda"
            }
        },
        "Test":
        {
            "required":["Id_test","Domande"],
            "properties":
            {
                "Id_test":{"type":"string"},
                "Domande":{"$ref":"#/definitions/ArrayDiDomande"}
            }
        },
        "BaseTest":
        {
            "required":["Id_test","Id_topic","Skill"],
            "properties":
            {
                "Id_test":{"type":"string"},
                "Id_topic":{"type":"integer"},
                "Skill":{"type":"integer"}
            }
        },
        "ArrayDiBaseTest":
        {
            "type":"array",
            
            "items":
            {
                "$ref":"#/definitions/BaseTest"
            }
        },
        "IdMateria":
        {
            "required":["IdMateria"],
            "properties":{"IdMateria":{"type":"string"}}
        },
        "Materia":
        {
            "required":["IdMateria","Nome"],
            "properties":{"IdMateria":{"type":"string"},"Nome":{"type":"string"}}
        },
        "ArrayDiMaterie":
        {
            "type":"array",
            "items":
            {
                "$ref":"#/definitions/Materia"
            }
        },
        "ArrayDiTopicSkill":
        {
            "type":"array",
            "items":
            {
                "$ref":"#/definitions/TopicSkillUtente"
            }
        }
        

    },
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"]
  }